Creating an Android Library
Creating an Android Library containing an Activity that will display a joke passed to it as an intent extra. Wire up project dependencies so that the button can now pass the joke from the Java Library to the Android Library.

FYI, the target version of your Android library and app MUST be thesame