Added a Paid and Free Flavor
Added free and paid product flavors to the app. Removed the ad and ad dependency from the paid flavor.

FYI, the fragment tht displays the ad was removed from the 'main' and added to the flavors with different functions. Free version retained the ads displayed by 'main' and paid flavor shows no ads